// frontend/src/lib/api/model/permissionGrantResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * dataing
 * Autonomous Data Quality Investigation
 * OpenAPI spec version: 2.0.0
 */
import type { PermissionGrantResponseDataSourceId } from "./permissionGrantResponseDataSourceId";
import type { PermissionGrantResponseGranteeId } from "./permissionGrantResponseGranteeId";
import type { PermissionGrantResponseResourceId } from "./permissionGrantResponseResourceId";
import type { PermissionGrantResponseTagId } from "./permissionGrantResponseTagId";

/**
 * Permission grant response.
 */
export interface PermissionGrantResponse {
  access_type: string;
  data_source_id: PermissionGrantResponseDataSourceId;
  grantee_id: PermissionGrantResponseGranteeId;
  grantee_type: string;
  id: string;
  permission: string;
  resource_id: PermissionGrantResponseResourceId;
  resource_type: string;
  tag_id: PermissionGrantResponseTagId;
}